requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg==3.1.14
pydantic==2.5.0
langchain==0.1.5
langchain-huggingface==0.0.1
langchain-core==0.1.7
langchain-groq==0.1.0
langchain-openai==0.0.7
langchain-google-genai==0.0.11
huggingface-hub==0.19.4
python-dotenv==1.0.0
